backend/api/requirements.txt
fastapi
uvicorn[standard]
httpx
asyncpg
python-dotenv
pyjwt[crypto]
